remove C# 8 code
set charactersLimit for string writer to 100 000
move pool classes into separate files inside Elastic.Apm.Helpers namespace
